+++ project_run/apps/athletes/serializers.py
@@ -1,11 +1,11 @@
-from rest_framework.serializers import ModelSerializer, IntegerField
+from rest_framework.serializers import ModelSerializer, IntegerField, FloatField
 
 from project_run.apps.athletes.models import AthleteInfo, validate_weight
 
 
 class AthleteInfoSerializer(ModelSerializer):
     user_id = IntegerField(read_only=True)
-    goals = IntegerField(required=False)
+    goals = FloatField(required=False)
     weight = IntegerField(validators=[validate_weight], required=False)
